Le scopecreep de l'infini : - AvisController avec post et delete - PhotoAvisController avec post (delete fait par avis) - Ajout de ModelState dans les put - Ajout de sécurité dans historique et cb - GET inutiles mis en commentaire

+++ SAE401-API/Controllers/CartebancaireController.cs
@@ -19,35 +19,36 @@ namespace SAE401_API.Controllers
             dataRepository = datarepo;
         }
 
-        [HttpGet("{idcartebancaire}")]
-        public async Task<ActionResult<CartebancaireDTO>> GetCartebancaireByIdAsync(int idcartebancaire)
-        {
-            var cartebancaire = await dataRepository.GetCartebancaireByIdAsync(idcartebancaire);
 
-            if (cartebancaire.Value == null)
+        // GET: api/Cartebancaire/GetAllCartebancaireByClient/idclient
+        [HttpGet("[action]/{idclient}")]
+        [Authorize()]
+        public async Task<ActionResult<IEnumerable<Cartebancaire>>> GetAllCartebancaireByClient(int idclient)
+        {
+            var identity = HttpContext.User.Identity as ClaimsIdentity;
+            if (identity == null || identity.FindFirst("id").Value != idclient.ToString())
             {
-                return NotFound();
+                return Forbid();
             }
 
-            var cartebancaireDTO = new CartebancaireDTO
-            {
-                Idcartebancaire = cartebancaire.Value.Idcartebancaire,
-                Idclient = cartebancaire.Value.Idclient,
-                Nomcartebancaire = cartebancaire.Value.Nomcartebancaire,
-                Dateenregistement = cartebancaire.Value.Dateenregistement,
-                Numcartebancaire = cartebancaire.Value.Numcartebancaire,
-                Dateexpirationcarte = cartebancaire.Value.Dateexpirationcarte
-            };
-
-            return cartebancaireDTO;
+            return await dataRepository.GetAllCartebancaireByClientAsync(idclient);
         }
+        
 
+        // POST: api/Cartebancaire
         [HttpPost]
+        [Authorize()]
         public async Task<ActionResult> PostCartebancaire([FromBody] CartebancaireDTO cartebancaireDTO)
         {
             if (!ModelState.IsValid)
             {
                 return BadRequest(ModelState);
+            }
+
+            var identity = HttpContext.User.Identity as ClaimsIdentity;
+            if (identity == null || identity.FindFirst("id").Value != cartebancaireDTO.Idclient.ToString())
+            {
+                return Forbid();
             }
 
             var cartebancaire = new Cartebancaire
@@ -63,9 +64,22 @@ namespace SAE401_API.Controllers
             return NoContent();
         }
 
+        // PUT: api/Cartebancaire/id
         [HttpPut("{idcartebancaire}")]
+        [Authorize()]
         public async Task<IActionResult> PutCartebancaire(int idcartebancaire, [FromBody] CartebancaireDTO cartebancaireDTO)
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
+            var identity = HttpContext.User.Identity as ClaimsIdentity;
+            if (identity == null || identity.FindFirst("id").Value != cartebancaireDTO.Idclient.ToString())
+            {
+                return Forbid();
+            }
+
             if (idcartebancaire != cartebancaireDTO.Idcartebancaire)
             {
                 return BadRequest("Les paramètres ne correspondent pas.");
@@ -91,13 +105,21 @@ namespace SAE401_API.Controllers
             return NoContent();
         }
 
+        // DELETE: api/Cartebancaire/id
         [HttpDelete("{idcartebancaire}")]
+        [Authorize()]
         public async Task<IActionResult> DeleteCartebancaire(int idcartebancaire)
         {
             var cartebancaire = await dataRepository.GetCartebancaireByIdAsync(idcartebancaire);
             if (cartebancaire.Value == null)
             {
                 return NotFound();
+            }
+
+            var identity = HttpContext.User.Identity as ClaimsIdentity;
+            if (identity == null || identity.FindFirst("id").Value != cartebancaire.Value.Idclient.ToString())
+            {
+                return Forbid();
             }
 
             await dataRepository.DeleteCartebancaireAsync(cartebancaire.Value);
